+++ include/cask/Scheduler.hpp
@@ -56,6 +56,16 @@ public:
     void submit(const std::function<void()>& task);
 
     /**
+     * Submit several tasks at once to the the thread pool. The order
+     * these tasks will be taken up and executed is undefined. Each
+     * task will execute after an indeterminite amount of time as
+     * resource free to perform the individual task.
+     * 
+     * @param tasks The vector of tasks to submit in-bulk.
+     */
+    void submitBulk(const std::vector<std::function<void()>>& tasks);
+
+    /**
      * Submit a task to the pool after _at least_ the given amount
      * of time has passed.
      * 
@@ -64,17 +74,26 @@ public:
      * @param task The task the submit after the wait time has elapsed.
      */
     void submitAfter(int64_t milliseconds, const std::function<void()>& task);
+
+    /**
+     * Check if the scheduler is currently idle - meaning all threads are
+     * currently waiting for tasks to execute.
+     * 
+     * @return true if the scheduler is idle.
+     */
+    bool isIdle() const;
 private:
     bool running;
 
     std::mutex readyQueueMutex;
     std::condition_variable dataInQueue;
     std::queue<std::function<void()>> readyQueue;
+    std::atomic_size_t idleThreads;
     std::mutex timerMutex;
     std::map<int64_t,std::vector<std::function<void()>>> timers;
     std::vector<std::thread> runThreads;
     std::thread timerThread;
-    std::atomic_long ticks;
+    int64_t ticks;
 
     void run();
     void timer();