+++ src/cask/Scheduler.cpp
@@ -19,10 +19,12 @@ Scheduler::Scheduler(int poolSize)
     , readyQueueMutex()
     , dataInQueue()
     , readyQueue()
+    , idleThreads(poolSize)
     , timerMutex()
     , timers()
     , runThreads()
     , timerThread()
+    , ticks(0)
 {
     for(int i =0; i < poolSize; i++) {
         std::thread poolThread(std::bind(&Scheduler::run, this));
@@ -54,27 +56,50 @@ void Scheduler::submit(const std::function<void()>& task) {
     dataInQueue.notify_one();
 }
 
-void Scheduler::submitAfter(int64_t milliseconds, const std::function<void()>& task) {
-    int64_t executionTick = ticks.load() + milliseconds;
-    {
-        std::lock_guard guard(timerMutex);
-        auto tasks = timers.find(executionTick);
-        if(tasks == timers.end()) {
-            std::vector<std::function<void()>> taskVector = {task};
-            timers[executionTick] = taskVector;
-        } else {
-            tasks->second.push_back(task);
-        }
+void Scheduler::submitBulk(const std::vector<std::function<void()>>& tasks) {
+    std::lock_guard guard(readyQueueMutex);
+    for(auto& task: tasks) {
+        readyQueue.emplace(task);
+        dataInQueue.notify_one();
     }
+}
+
+void Scheduler::submitAfter(int64_t milliseconds, const std::function<void()>& task) {
+    std::lock_guard guard(timerMutex);
+    int64_t executionTick = ticks + milliseconds;
+    auto tasks = timers.find(executionTick);
+    if(tasks == timers.end()) {
+        std::vector<std::function<void()>> taskVector = {task};
+        timers[executionTick] = taskVector;
+    } else {
+        tasks->second.push_back(task);
+    }
+}
+
+bool Scheduler::isIdle() const {
+    return idleThreads.load() == runThreads.size() && readyQueue.empty();
 }
 
 void Scheduler::run() {
     std::unique_lock<std::mutex> readyQueueLock(readyQueueMutex, std::defer_lock);
     std::chrono::milliseconds max_wait_time(10);
     std::function<void()> task;
+    bool idling = true;
+
     while(running) {
         readyQueueLock.lock();
+
+        if(!idling && readyQueue.empty()) {
+            idling = true;
+            idleThreads++;
+        }
+
         if(dataInQueue.wait_for(readyQueueLock, max_wait_time, [this](){return !readyQueue.empty(); })) {
+            if(idling) {
+                idling = false;
+                idleThreads--;
+            }
+            
             task = readyQueue.front();
             readyQueue.pop();
             readyQueueLock.unlock();
@@ -86,18 +111,28 @@ void Scheduler::run() {
 }
 
 void Scheduler::timer() {
+    const static std::chrono::milliseconds sleep_time(10);
+
     while(running) {
-        std::this_thread::sleep_for(std::chrono::milliseconds(1));
-        int64_t currentTick = ticks.fetch_add(1);
+        auto before = std::chrono::high_resolution_clock::now();
+        std::this_thread::sleep_for(sleep_time);
+        auto after = std::chrono::high_resolution_clock::now();
+        
+        auto delta = after - before;
+        auto milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(delta).count();
 
         {
             std::lock_guard guard(timerMutex);
-            auto tasks = timers.find(currentTick);
-            if(tasks != timers.end()) {
-                for(auto& task: tasks->second) {
-                    submit(task);
+
+            int64_t previous_ticks = ticks;
+            ticks += milliseconds;
+
+            for(int64_t i = previous_ticks; i <= ticks; i++) {
+                auto tasks = timers.find(i);
+                if(tasks != timers.end()) {
+                    submitBulk(tasks->second);
+                    timers.erase(i);
                 }
-                timers.erase(currentTick);
             }
         }
     }