+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All feature additions, significant bug fixes, and API changes will be documented
 in this file. This project follows [semantic versioning](https://semver.org/).
 
+## 7.3
+
+- Make `Scheduler` evaluate timers far less often (every 10ms rather than 1ms).
+  Enhance the timer handler to better deal with scew where the OS does not
+  schedule the timer thread on exact scheduled boundaries.
+- Add `Scheduler::submitBulk` as an optimization for clients which wawnt to
+  submit several items to the scheduler in a single operation. 
+- Add `Scheduler::isIdle` to allow users to check if the scheduler is currently
+  idle (that is - not executing tasks and nothing in the ready queue).
+
 ## 7.2
 
 - Add the `Queue` type as a generalizaton of `MVar` but for mailboxes